Add spark listener to public events to the Yarn AHS.
The listener depends on Hadoop 2.4, so its compilation is
only enabled if a compatible version of Hadoop is being
used. It can be explicitly disabled by the user, just in
case.

The Yarn scheduler backend will automatically instantiate
the listener if (i) it's available and (ii) the Yarn
configuration contains the address of the history server.
There is a new config option ("spark.yarn.history.enabled")
that allows the user to disable using the history server
when starting a job.

Build is automatically enabled in sbt if Yarn is enabled and
the Hadoop version is current enough. A separate profile
(yarn-history) needs to be enabled for the code to be built
using maven.